Refactor `prompt` directory to `ai-function`
This change restructures the codebase by renaming and relocating
files from the `prompt` directory to the `ai-function` directory.
This includes moving functions like `createAIFunction` and
`createAIRunner`, along with their corresponding types and tests.
