+++ examples/ai-function.ts
@@ -1,7 +1,7 @@
 import 'dotenv/config';
 
+import { createAIFunction } from '@dexaai/dexter/ai-function';
 import { ChatModel, type Msg, MsgUtil } from '@dexaai/dexter/model';
-import { createAIFunction } from '@dexaai/dexter/prompt';
 import { z } from 'zod';
 
 /**
