+++ examples/ai-runner-with-error.ts
@@ -1,7 +1,7 @@
 import 'dotenv/config';
 
+import { createAIFunction, createAIRunner } from '@dexaai/dexter/ai-function';
 import { ChatModel, MsgUtil } from '@dexaai/dexter/model';
-import { createAIFunction, createAIRunner } from '@dexaai/dexter/prompt';
 import { z } from 'zod';
 
 /** Get the weather for a given location. */
