+++ package.json
@@ -15,9 +15,9 @@
   },
   "exports": {
     "./package.json": "./package.json",
-    "./prompt": {
-      "types": "./dist/prompt/index.d.ts",
-      "import": "./dist/prompt/index.js"
+    "./ai-function": {
+      "types": "./dist/ai-function/index.d.ts",
+      "import": "./dist/ai-function/index.js"
     },
     "./model": {
       "types": "./dist/model/index.d.ts",
