+++ src/model/utils/extract-json.ts
