+++ src/model/utils/extract-zod-object.test.ts
