+++ src/model/utils/extract-zod-object.ts
