+++ src/model/utils/zod-to-json-schema.ts
@@ -1,7 +1,7 @@
 import { type z } from 'zod';
 import { zodToJsonSchema as zodToJsonSchemaImpl } from 'zod-to-json-schema';
 
-import { omit } from '../../model/utils/helpers.js';
+import { omit } from './helpers.js';
 
 /** Generate a JSON Schema from a Zod schema. */
 export function zodToJsonSchema(schema: z.ZodType): Record<string, unknown> {
