--- src/prompt/functions/stringify-for-model.ts
@@ -1,18 +1,0 @@
-import { type Jsonifiable } from 'type-fest';
-
-/**
- * Stringifies a JSON value in a way that's optimized for use with LLM prompts.
- *
- * This is intended to be used with `function` and `tool` arguments and responses.
- */
-export function stringifyForModel(jsonObject: Jsonifiable | void): string {
-  if (jsonObject === undefined) {
-    return '';
-  }
-
-  if (typeof jsonObject === 'string') {
-    return jsonObject;
-  }
-
-  return JSON.stringify(jsonObject, null, 0);
-}
